Study timer database integration
I am refactoring the code to make it cleaner as well as integrate a sqlite3 database to improve functionality. The updated code will take the user input for the subject name and create a table for that subject. The date, number of sessions, and time spent studying will all be saved to the appropriate table.