Code refactor + Plus session overview
Outer while statement changed from 'while option != 3' to 'while True' since there was already a conditional statement to break out of the loop when option 3 is selected. Once a study session is completed, it will give an overview of the number of study blocks and hours spent studying for that study session.